moved AggregationState to aggregation_properties.py
